requirements.txt
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.21
Werkzeug==2.2.2
pytest==7.1.2
requests
python-dotenv